% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{add_channel}
\alias{add_channel}
\title{Get and set job channels}
\usage{
add_channel(id, channel, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{channel}{A character string containing the name of a channel, as returned by \code{\link{list_channels}}.}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A logical \code{TRUE}, or an error.
}
\description{
Get, set, and replace the \dQuote{tags} for a job
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- create_job(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# list available channels
ch <- list_channels(j)
ch

# add new channel
add_channel(j, ch[1])

# remove a channel
remove_channel(j, ch[1])
}
}
\seealso{
\code{\link{create_job}}
}

