% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{Uploads data to a CrowdFlower Job}
\usage{
add_data(id, data, verbose = TRUE, ...)
}
\arguments{
\item{id}{ID for job to be updated.}

\item{data}{A filename or data.frame with data to be uploaded.}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A character string containing the job ID, invisibly.
}
\description{
\code{add_data} uploads data from a data.frame to a CrowdFlower Job
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- create_job(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

d <- data.frame(variable = 1:3)
add_data(id = j, data = d)
}
}
\references{
\href{https://success.crowdflower.com/hc/en-us/articles/202703425-CrowdFlower-API-Requests-Guide#data}{Crowdflower API documentation}
}
\seealso{
\code{\link{update_job}}, \code{\link{get_status}}, \code{\link{get_results}}
}

