% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{APIcall}
\alias{APIcall}
\title{Execute API call to CrowdFlower API.}
\usage{
APIcall(endpoint, params = NULL, body = NULL, type = "GET",
  key = Sys.getenv("CROWDFLOWER_API_KEY"),
  base_url = "https://api.crowdflower.com/v1/")
}
\arguments{
\item{endpoint}{endpoint of API request.}

\item{params}{additional text to add to request URL.}

\item{body}{additional parameters to add to request body.}

\item{type}{type of requests (GET, POST, PUT).}

\item{key}{A character string containing a Crowdflower API key.}

\item{base_url}{Base URL of API request.}
}
\value{
A character string containing the JSON response from the Crowdflower API.
}
\description{
\code{APIcall} connects to the CrowdFlower API and executes an
API request.
}

