% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_account.R
\name{cf_account}
\alias{cf_account}
\title{Retrieve account information}
\usage{
cf_account(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A list containing account details for the authenticated user.
}
\description{
\code{getAccount} retrieves the CrowdFlower account information
related to the API key for the authenticated user
}
\examples{
\dontrun{
getAccount()
}
}
\references{
\href{https://success.crowdflower.com/hc/en-us/articles/202703425-CrowdFlower-API-Requests-Guide#account_information}{Crowdflower API documentation}
}
\keyword{account}

