% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_jobs_list.R
\name{cf_jobs_list}
\alias{cf_jobs_list}
\title{Retrieve list of all jobs.}
\usage{
cf_jobs_list(page = NULL, ...)
}
\arguments{
\item{page}{A vector of integers specifying which page(s) of results to return. A page contains up to 10 jobs. The default (\code{NULL}) is to return all jobs.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A data.frame containing details of all jobs. The \code{id} column provides the Crowdflower Job ID for each job.
}
\description{
\code{cf_jobs_list} queries for all jobs under the account related to the API key for the authenticated user.
}
\details{
Use \code{\link{job_get}} to return a single job object as a list.
}
\examples{
\dontrun{
# return first page of jobs
cf_jobs_list(page = 1)

# return all jobs
cf_jobs_list()
}
}
\references{
\href{https://success.crowdflower.com/hc/en-us/articles/202703425-CrowdFlower-API-Requests-Guide#account_information}{Crowdflower API documentation}
}
\seealso{
\code{\link{cf_account}}, \code{\link{job_get}}
}
\keyword{account,}
\keyword{jobs}

