% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_query.R
\name{cf_query}
\alias{cf_query}
\title{Execute a Crowdflower API call}
\usage{
cf_query(endpoint, query = NULL, body = NULL, type = "GET",
  parse = TRUE, key = Sys.getenv("CROWDFLOWER_API_KEY"),
  base_url = "https://api.crowdflower.com/v1/", ...)
}
\arguments{
\item{endpoint}{A character string specifying the endpoint of API request relative to \code{base_url}}

\item{query}{A list containing named URL query parameters to add to request URL.}

\item{body}{Additional parameters to add to request body.}

\item{type}{A character string specifying the type of HTTP request (\dQuote{GET}, \dQuote{POST}, \dQuote{PUT}, etc.).}

\item{parse}{A logical indicating whether to parse the response (\code{TRUE}, the default) or return a raw vector (\code{FALSE}).}

\item{key}{A character string containing a Crowdflower API key.}

\item{base_url}{A character string specifying the base URL of API request.}

\item{...}{Additional arguments passed to an HTTP request function (e.g, \code{\link[httr]{GET}}, \code{\link[httr]{POST}}, \code{\link[httr]{PUT}}).}
}
\value{
A character string containing the content of the HTTP request.
}
\description{
Connects to the CrowdFlower API and executes an API request, returning the content of the HTTP response.
}
