% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{addChannel}
\alias{addChannel}
\alias{removeChannel}
\title{Get and set job channels}
\usage{
addChannel(id, channel, verbose = TRUE, ...)

removeChannel(id, channel, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{channel}{A character string containing the name of a channel, as returned by \code{\link{listChannels}}.}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{crowdflowerAPIQuery}}.}
}
\value{
A logical \code{TRUE}, or an error.
}
\description{
Get, set, and replace the \dQuote{tags} for a job
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- createJob(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# list available channels
ch <- listChannels(j)
ch

# add new channel
addChannel(j, ch[1])

# remove a channel
removeChannel(j, ch[1])
}

}
\seealso{
\code{\link{createJob}}
}

