% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_job.R
\name{copy_job}
\alias{copy_job}
\title{Copy a job and, optionally, its data to a new job ID.}
\usage{
copy_job(id, rows = TRUE, gold = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{rows}{A logical indicating whether to copy rows from the original job.}

\item{gold}{A logical indicating whether to copy \emph{only} gold questions from the original job. Ignored if \code{rows = TRUE}.}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A character string containing the job ID, invisibly.
}
\description{
\code{copy_job} copies an existing job to a new job ID.
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", 
                  package = "crowdflower")
f2 <- system.file("templates/cml1.xml", 
                  package = "crowdflower")
j1 <- create_job(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# copy job w/rows
j2 <- copy_job(id = j1)

# copy job w/only gold questions
j2 <- copy_job(id = j1, rows = FALSE, gold = TRUE)

# copy job w/o any rows
j2 <- copy_job(id = j1, rows = FALSE)
}
}
\seealso{
\code{\link{create_job}}, \code{\link{update_job}}
}
\keyword{jobs}

