% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_job.R
\name{create_job}
\alias{create_job}
\title{Creates a new job}
\usage{
create_job(title = NULL, instructions = NULL, cml = NULL,
  verbose = TRUE, ...)
}
\arguments{
\item{title}{A character string containing the title for a new job.}

\item{instructions}{A character string containing instructions for a new job.}

\item{cml}{Layout for new job.}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
An integer specify the job ID.
}
\description{
\code{create_job} creates a new CrowdFlower Job.
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- create_job(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))
}
}
\references{
\href{https://success.crowdflower.com/hc/en-us/articles/202703425-CrowdFlower-API-Requests-Guide}{Crowdflower API documentation}
}
\seealso{
\code{\link{update_job}}, \code{\link{add_data}}, \code{\link{launch_job}}, \code{\link{get_status}}, \code{\link{get_results}}
}

