% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_contributor.R
\name{flag_contributor}
\alias{flag_contributor}
\alias{reject_contributor}
\alias{unflag_contributor}
\title{Flag/unflag/reject Contributors}
\usage{
flag_contributor(id, worker, reason, ...)

unflag_contributor(id, worker, reason, ...)

reject_contributor(id, worker, reason, ...)
}
\arguments{
\item{id}{A character string containing an ID for job to be updated.}

\item{worker}{A character string containing a worker/contributor ID.}

\item{reason}{A character string containing a reason for flagging/unflagging/rejecting.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A character string containing the job ID, invisibly.
}
\description{
Flag/unflag or reject a contributor
}
\examples{
\dontrun{
# Flag contributor
flag_contributor(j, worker, "Looks problematic")

# Unflag contributor
unflag_contributor(j, worker, "Actually, all is well")

# Reject contributor
reject_contributor(j, worker, "This is unacceptable.")
}

}
\seealso{
\code{\link{create_job}}
}
\keyword{contributors}

