% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_contributor.R
\name{flagContributor}
\alias{flagContributor}
\alias{unflagContributor}
\title{Flag Contributors}
\usage{
flagContributor(id, worker, reason, ...)

unflagContributor(id, worker, reason, ...)
}
\arguments{
\item{id}{A character string containing an ID for job to be updated.}

\item{worker}{A character string containing a worker/contributor ID.}

\item{reason}{A character string containing a reason for flagging/unflagging/rejecting.}

\item{...}{Additional arguments passed to \code{\link{crowdflowerAPIQuery}}.}
}
\value{
A character string containing the job ID, invisibly.
}
\description{
Flag/unflag or reject a contributor
}
\examples{
\dontrun{
# Flag contributor
flagContributor(j, worker, "Looks problematic")

# Unflag contributor
unflagContributor(j, worker, "Actually, all is well")

# Reject contributor
rejectContributor(j, worker, "This is unacceptable.")
}

}
\seealso{
\code{\link{createJob}}
}

