% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legend.R
\name{get_job_legend}
\alias{get_job_legend}
\title{Get Job Legend}
\usage{
get_job_legend(id, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A list
}
\description{
Get the \dQuote{legend} (description) for a job
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- create_job(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# get the legend
get_job_legend(id, title = "New Title")
}
}
\seealso{
\code{\link{create_job}}
}
\keyword{jobs}

