% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jobs.R
\name{get_jobs}
\alias{get_jobs}
\title{Retrieve list of all jobs.}
\usage{
get_jobs(page = NULL, ...)
}
\arguments{
\item{page}{A vector of integers specifying which page(s) of results to return. A page contains up to 10 jobs. The default (\code{NULL}) is to return all jobs.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A data.frame containing details of all jobs. The \code{id} 
column provides the Crowdflower Job ID for each job.
}
\description{
\code{get_jobs} queries for all jobs under the account related to the API key for the authenticated user.
}
\examples{
\dontrun{
# return first page of jobs
get_jobs(page = 1)

# return all jobs
get_jobs()
}
}
\references{
\href{https://success.crowdflower.com/hc/en-us/articles/202703425-CrowdFlower-API-Requests-Guide#account_information}{Crowdflower API documentation}
}
\seealso{
\code{\link{get_account}}
}

