% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\alias{regenerate_report}
\title{Generate and retrieve job results}
\usage{
get_results(id, report_type = c("full", "aggregated", "json", "gold_report",
  "workset", "source"), verbose = TRUE, ...)

regenerate_report(id, report_type = c("full", "aggregated", "json",
  "gold_report", "workset", "source"), verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{report_type}{Type of report}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\description{
\code{get_results} regenerates one of six types of reports within
a job:
- \code{full}: Returns the Full report containing every judgment.
- \code{aggregated}: Returns the Aggregated report containing the 
aggregated response for each row.
- \code{json}: Returns the JSON report containing the aggregated 
response, as well as the individual judgments, for the first 50 rows.
- \code{gold_report}: Returns the Test Question report.
- \code{workset}: Returns the Contributor report.
- \code{source}: Returns a CSV of the source data uploaded to the job.
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- create_job(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# get results for job
get_results(id = j, report_type = "")
}
}
\references{
\href{https://success.crowdflower.com/hc/en-us/articles/202703425-CrowdFlower-API-Requests-Guide#get_results}{Crowdflower API documentation}
}
\seealso{
\code{\link{get_account}}
}

