% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rows.R
\name{get_rows}
\alias{get_rows}
\title{Retrieve rows with judgment information for a job.}
\usage{
get_rows(id, n = Inf, type = "aggregated", verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{n}{Number of rows to return.}

\item{type}{How responses should be aggregated ("aggregate", to take aggregated responses; or "full" for all responses)}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A data.frame containing judgment information
}
\description{
\code{get_rows} retrieves rows with judgment information for the job ID indicated in the request.
}
\examples{
\dontrun{
get_rows('jobid')
}
}
\seealso{
\code{\link{get_status}}, \code{\link{get_results}}
}

