% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status.R
\name{get_status}
\alias{get_status}
\title{Retrieve the status of a job}
\usage{
get_status(id, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
Details about a job.
}
\description{
\code{get_status} gets the status of the job identified by the id parameter of the request.
}
\examples{
\dontrun{
j <- create_job(title = 'Job Title', 
               instructions = 'Some instructions')

# get status
get_status(j)
}
}
\seealso{
\code{\link{get_rows}}, \code{\link{get_results}}
}

