% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-account.R
\name{getAccount}
\alias{getAccount}
\title{Retrieve account information}
\usage{
getAccount()
}
\description{
\code{getAccount} retrieves the CrowdFlower account information
related to the API key for the authenticated user
}

