% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legend.R
\name{getJobLegend}
\alias{getJobLegend}
\title{Get Job Legend}
\usage{
getJobLegend(id, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{...}{Additional arguments passed to \code{\link{APIcall}}.}
}
\value{
A list
}
\description{
Get the \dQuote{legend} for a job
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- createJob(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# get the legend
getJobLegend(id, title = "New Title")

}

}
\seealso{
\code{\link{createJob}}
}

