% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_results.R
\name{getResults}
\alias{getResults}
\title{Generate and retrieve job results}
\usage{
getResults(id, report_type, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{report_type}{Type of report}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{APIcall}}.}
}
\description{
\code{getResults} regenerates one of six types of reports within
a job:
- \code{full}: Returns the Full report containing every judgment.
- \code{aggregated}: Returns the Aggregated report containing the
aggregated response for each row.
- \code{json}: Returns the JSON report containing the aggregated
response, as well as the individual judgments, for the first 50 rows.
- \code{gold_report}: Returns the Test Question report.
- \code{workset}: Returns the Contributor report.
- \code{source}: Returns a CSV of the source data uploaded to the job.
}
\examples{
\dontrun{
getResults(id = "jobid", report_type = "")
}
}
\references{
\href{https://success.crowdflower.com/hc/en-us/articles/202703425-CrowdFlower-API-Requests-Guide#get_results}{Crowdflower API documentation}
}
\seealso{
\code{\link{getAccount}}
}

