% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_rows.R
\name{getRows}
\alias{getRows}
\title{Retrieve rows with judgment information for a job.}
\usage{
getRows(id, n = Inf, type = "aggregated")
}
\arguments{
\item{id}{ID for job.}

\item{n}{Number of rows to return.}

\item{type}{How responses should be aggregated ("aggregate", to take
aggregated responses; or "full" for all responses)}
}
\description{
\code{getRows} retrieves rows with judgment information for the
job ID indicated in the request.
}

