% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status.R
\name{getStatus}
\alias{getStatus}
\title{Retrieve the status of a job}
\usage{
getStatus(id, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{...}{Additional arguments passed to \code{\link{APIcall}}.}
}
\value{
Details about a job.
}
\description{
\code{getStatus} gets the status of the job identified by the 
id parameter of the request.
}
\examples{
\dontrun{
j <- createJob(title = 'Job Title', 
               instructions = 'Some instructions')

# get status
getStatus(j)
}

}
\seealso{
\code{\link{getRows}}, \code{\link{getResults}}
}

