% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_add_data.R
\name{job_add_data}
\alias{job_add_data}
\title{Uploads data to a CrowdFlower Job}
\usage{
job_add_data(id, data, force = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{id}{ID for job to be updated.}

\item{data}{A filename or data.frame with data to be uploaded.}

\item{force}{A logical indicating whether to \emph{force} the upload even when column names do not match existing headers.}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A character string containing the job ID, invisibly.
}
\description{
\code{job_add_data} uploads data from a data.frame to a CrowdFlower Job
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- job_create(title = "Job Title", 
                instructions = readChar(f1, nchars = 1e8L),
                cml = readChar(f2, nchars = 1e8L))

# add data
d <- data.frame(variable = 1:3)
job_add_data(id = j, data = d)

# launch job
job_launch(id = j)

# get results for job
report_regenerate(id = j, report_type = "full")
report_get(id = j, report_type = "full")

# delete job
job_delete(j)
}
}
\references{
\href{https://success.crowdflower.com/hc/en-us/articles/202703425-CrowdFlower-API-Requests-Guide#data}{Crowdflower API documentation}
}
\seealso{
\code{\link{job_update}}, \code{\link{job_status}}, \code{\link{report_get}}
}
\keyword{data}
\keyword{jobs}
