% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_contributor.R
\name{job_contributor_flag}
\alias{job_contributor_flag}
\alias{job_contributor_reject}
\alias{job_contributor_unflag}
\title{Flag/unflag/reject Contributors}
\usage{
job_contributor_flag(id, contributor, reason, ...)

job_contributor_unflag(id, contributor, reason, ...)

job_contributor_reject(id, contributor, reason, ...)
}
\arguments{
\item{id}{A character string containing an ID for job to be updated.}

\item{contributor}{A character string containing a contributor/contributor ID.}

\item{reason}{A character string containing a reason for flagging/unflagging/rejecting.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A character string containing the job ID, invisibly.
}
\description{
Flag/unflag or reject a contributor
}
\examples{
\dontrun{
# Flag contributor
job_contributor_flag(j, contributor, "Looks problematic")

# Unflag contributor
job_contributor_unflag(j, contributor, "Actually, all is well")

# Reject contributor
job_contributor_reject(j, contributor, "This is unacceptable.")
}

}
\seealso{
\code{\link{job_create}}
}
\keyword{contributors}

