% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_uploaded_gold.R
\name{job_convert_gold}
\alias{job_convert_gold}
\title{convert uploaded gold for a job}
\usage{
job_convert_gold(id)
}
\description{
Convert to test questions all rows where column "_golden" is set to "true".
Corresponds to the "Convert Uploaded Test Questions" button on the Data page
of the web interface.
}

