% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_create.R
\name{job_create}
\alias{job_create}
\title{Creates a new job}
\usage{
job_create(title = NULL, instructions = NULL, cml = NULL,
  verbose = TRUE, ...)
}
\arguments{
\item{title}{A character string containing the title for a new job.}

\item{instructions}{A character string containing instructions for a new job.}

\item{cml}{Layout for new job.}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
An integer specify the job ID.
}
\description{
\code{job_create} creates a new CrowdFlower Job.
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- job_create(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))
}
}
\references{
\href{https://success.crowdflower.com/hc/en-us/articles/202703425-CrowdFlower-API-Requests-Guide}{Crowdflower API documentation}
}
\seealso{
\code{\link{job_update}}, \code{\link{job_add_data}}, \code{\link{job_launch}}, \code{\link{job_status}}, \code{\link{report_get}}
}
\keyword{jobs}

