% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_delete.R
\name{job_delete}
\alias{job_delete}
\title{Delete a single job}
\usage{
job_delete(id, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A logical.
}
\description{
Delete a single Crowdflower job
}
\keyword{jobs}

