% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_get.R
\name{job_get}
\alias{job_get}
\title{Get a single job}
\usage{
job_get(id, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A list containing details of all jobs. The \code{id} field provides the Crowdflower Job ID for each job.
}
\description{
Get a single Crowdflower job
}
\seealso{
\code{\link{cf_jobs_list}}, \code{\link{cf_account}}
}
\keyword{jobs}

