% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_control.R
\name{job_launch}
\alias{job_cancel}
\alias{job_launch}
\alias{job_pause}
\alias{job_resume}
\title{Control a job}
\usage{
job_launch(id, channel = "cf_internal", units = 100, verbose = FALSE, ...)

job_pause(id, verbose = FALSE, ...)

job_cancel(id, verbose = FALSE, ...)

job_resume(id, verbose = FALSE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{channel}{Either "on_demand" (for On Demand worforce), \dQuote{cf_internal} (for the Internal Channel) or "both"}

\item{units}{Units to be launched.}

\item{verbose}{A logical indicating whether to print details of the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
For \code{job_launch}, a list containing details of the job. Otherwise, a logical indicating the operation was successful, otherwise an error.
}
\description{
Launch, pause, resume, or cancel a job
}
\details{
\code{job_launch} launches a new job. \code{job_pause} pauses a currently running job. \code{job_resume} resumes a job. \code{job_cancel} cancels a job.
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- job_create(title = "Job Title", 
                instructions = readChar(f1, nchars = 1e8L),
                cml = readChar(f2, nchars = 1e8L))

# add data
d <- data.frame(variable = 1:3)
job_add_data(id = j, data = d)

# launch the job
job_launch(j)

# pause the job
job_pause(j)

# resume the job
job_resume(j)

# cancel the job
job_cancel(j)

# delete job
job_delete(j)
}
}
\seealso{
\code{\link{job_pause}}, \code{\link{job_cancel}}, \code{\link{job_resume}}
}
\keyword{jobs}

