% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_status.R
\name{job_status}
\alias{job_status}
\title{Retrieve the status of a job}
\usage{
job_status(id, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
Details about a job.
}
\description{
\code{job_status} gets the status of the job identified by the id parameter of the request.
}
\examples{
\dontrun{
j <- job_create(title = 'Job Title', 
                instructions = 'Some instructions')

# get status
job_status(j)

# delete job
job_delete(j)
}
}
\seealso{
\code{\link{job_get}}, \code{\link{results_get}}, \code{\link{report_get}}
}
\keyword{jobs}

