% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/control.R
\name{launchJob}
\alias{launchJob}
\title{Pause a job}
\usage{
launchJob(id, channel = "cf_internal", units = 100, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{channel}{Either "on_demand" (for On Demand worforce), "cf_internal"
(for the Internal Channel) or "both"}

\item{units}{Units to be launched.}

\item{...}{Additional arguments passed to \code{\link{APIcall}}.}
}
\value{
A character string containing the job ID, invisibly.
}
\description{
\code{launchJob} launches a new job.
}
\examples{
\dontrun{
launchJob('jobid')
}
}
\seealso{
\code{\link{pauseJob}}, \code{\link{cancelJob}}, \code{\link{resumeJob}}
}

