% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_channels.R
\name{listChannels}
\alias{listChannels}
\title{List available channels}
\usage{
listChannels(id, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{...}{Additional arguments passed to \code{\link{crowdflowerAPIQuery}}.}
}
\value{
A list.
}
\description{
Retrieve a list of available channels for a job
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- createJob(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# list available channels
ch <- listChannels(j)
ch

# add new channel
addChannel(j, ch[1])

# remove a channel
removeChannel(j, ch[1])
}

}
\seealso{
\code{\link{addChannel}}, \code{\link{removeChannel}}
}

