% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/control.R
\name{pauseJob}
\alias{pauseJob}
\title{Pause a job}
\usage{
pauseJob(id, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{...}{Additional arguments passed to \code{\link{APIcall}}.}
}
\value{
A character string contianing the job ID, invisibly.
}
\description{
\code{pauseJob} pauses a currently running job.
}
\examples{
\dontrun{
pauseJob('jobid')
}
}
\seealso{
\code{\link{cancelJob}}, \code{\link{resumeJob}}
}

