% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pay_contributor.R
\name{payBonus}
\alias{notifyContributor}
\alias{payBonus}
\title{Contact Contributors}
\usage{
payBonus(id, worker, amount, ...)

notifyContributor(id, worker, msg, ...)
}
\arguments{
\item{id}{A character string containing an ID for job to be updated.}

\item{worker}{A character string containing a worker/contributor ID.}

\item{amount}{A character string containing a bonus amount in U.S. cents.}

\item{...}{Additional arguments passed to \code{\link{crowdflowerAPIQuery}}.}

\item{msg}{A character string containing a message to be sent to the contributor.}
}
\value{
A character string containing the job ID, invisibly.
}
\description{
Contact a contributor via email or pay a contributor a bonus in U.S. cents
}
\examples{
\dontrun{
# Send message
notifyContributor(j, w, "Great work!")

# Pay bonus
payBonus(j, w, "0.01")
}

}
\seealso{
\code{\link{createJob}}
}

