% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\docType{data}
\name{Job}
\alias{Contributor}
\alias{Job}
\title{Job R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Job

Contributor
}
\description{
An R6 class for managing Crowdflower jobs using reference semantics
}
\details{
This is an experimental R6 interface for working with Crowdflower jobs
}
\seealso{
\code{\link{job_create}}
}
\keyword{datasets}

