% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/control.R
\name{resumeJob}
\alias{resumeJob}
\title{Pause a job}
\usage{
resumeJob(id, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{...}{Additional arguments passed to \code{\link{APIcall}}.}
}
\value{
A character string contianing the job ID, invisibly.
}
\description{
\code{resumeJob} resumes a currently paused job.
}
\examples{
\dontrun{
getStatus('jobid')
}
}
\seealso{
\code{\link{cancelJob}}, \code{\link{pauseJob}}
}

