% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{get_tags}
\alias{add_tags}
\alias{get_tags}
\alias{replace_tags}
\title{Get and set job tags}
\usage{
get_tags(id, ...)

add_tags(id, tags, verbose = TRUE, ...)

replace_tags(id, tags, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}

\item{tags}{For \code{add_tags}, a character vector specifying one or more tags to add. For \code{replace_tags}, the same but to \emph{replace} rather than \emph{add} the tags.}

\item{verbose}{A logical indicating whether to print additional information about the request.}
}
\value{
For \code{get_tags}, a character vector of tags. Otherwise, a logical \code{TRUE}, or an error.
}
\description{
Get, set, and replace the \dQuote{tags} for a job
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- create_job(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# get tags
get_tags(j)

# add new tag
add_tags(j, "textanalysis")

# replace tags
add_tags(j, c("foo", "bar"))
get_tags(j)
}
}
\seealso{
\code{\link{create_job}}
}

