% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_job.R
\name{update_job}
\alias{update_job}
\title{Updates job settings.}
\usage{
update_job(id, title = NULL, instructions = NULL, cml = NULL,
  payment_cents = NULL, units_per_assignment = NULL, auto_launch = NULL,
  verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job to be updated.}

\item{title}{New title.}

\item{instructions}{New set of instructions.}

\item{cml}{New layout for job.}

\item{payment_cents}{Amount, in U.S. cents, that contributors will be paid per task.}

\item{units_per_assignment}{Number of rows that will comprise one page within a job.}

\item{auto_launch}{A logical indicating whether to automatically launch rows as they are added.}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A character string containing the job ID, invisibly.
}
\description{
Updates settings for a given job.
}
\examples{
\dontrun{
# update job's title
update_job(id, title = "New Title")

# update job's payment
update_job(id, payment_cents = "0.1")

# update job's instructions and units
update_job(id, instructions = "blank", units_per_assignment = 3)
}
}
\seealso{
\code{\link{create_job}}
}

