% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_job.R
\name{updateJob}
\alias{updateJob}
\title{Updates settings for a given job.}
\usage{
updateJob(id, title = NULL, instructions = NULL, cml = NULL,
  payment_cents = NULL, units_per_assignment = NULL, ...)
}
\arguments{
\item{id}{A character string containing an ID for job to be updated.}

\item{title}{New title.}

\item{instructions}{New set of instructions.}

\item{cml}{New layout for job.}

\item{payment_cents}{Amount, in cents, that contributors will be paid per task.}

\item{units_per_assignment}{Number of rows that will comprise one page within a job.}

\item{...}{Additional arguments passed to \code{\link{APIcall}}.}
}
\value{
A character string containing the job ID, invisibly.
}
\description{
Updates settings for a given job.
}
\examples{
\dontrun{
# update job's title
updateJob(id, title = "New Title")
}
}
\seealso{
\code{\link{createJob}}
}

