% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{addData}
\alias{addData}
\title{Uploads data to a CrowdFlower Job}
\usage{
addData(id, data, verbose = TRUE, ...)
}
\arguments{
\item{id}{ID for job to be updated.}

\item{data}{A filename or data.frame with data to be uploaded.}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{crowdflowerAPIQuery}}.}
}
\value{
A character string containing the job ID, invisibly.
}
\description{
\code{addData} uploads data from a data.frame to a CrowdFlower Job
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- createJob(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

d <- data.frame(variable = 1:3)
addData(id = j, data = d)
}
}
\references{
\href{https://success.crowdflower.com/hc/en-us/articles/202703425-CrowdFlower-API-Requests-Guide#data}{Crowdflower API documentation}
}
\seealso{
\code{\link{updateJob}}, \code{\link{getStatus}}, \code{\link{getResults}}
}

