% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{APIcall}
\alias{APIcall}
\title{Execute API call to CrowdFlower API.}
\usage{
APIcall(endpoint, params = NULL, body = NULL, type = "GET",
  base_url = "https://api.crowdflower.com/v1/")
}
\arguments{
\item{endpoint}{endpoint of API request.}

\item{params}{additional text to add to request URL.}

\item{body}{additional parameters to add to request body.}

\item{type}{type of requests (GET, POST, PUT).}

\item{base_url}{Base URL of API request.}
}
\description{
\code{APIcall} connects to the CrowdFlower API and executes an
API request.
}

