% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{authCF}
\alias{authCF}
\title{Authenticate connection to CrowdFlower API.}
\usage{
authCF(api_key = NULL)
}
\arguments{
\item{api_key}{User's API key to connect to CrowdFlower's API, available in
\url{https://make.crowdflower.com/account/user}.}
}
\description{
\code{authCF} stores the user's API Key in a local environment
and checks that the connection was successful.
}
\seealso{
\url{https://success.crowdflower.com/hc/en-us/articles/202703445-CrowdFlower-API-Integrating-with-the-API#api_key}
}

