% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/control.R
\name{cancelJob}
\alias{cancelJob}
\title{Pause a job}
\usage{
cancelJob(id, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{...}{Additional arguments passed to \code{\link{APIcall}}.}
}
\value{
A character string contianing the job ID, invisibly.
}
\description{
\code{pauseJob} cancels a currently running job.
}
\examples{
\dontrun{
cancelJob('jobid')
}
}
\seealso{
\code{\link{pauseJob}}, \code{\link{resumeJob}}
}

