% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_account.R
\name{cf_account}
\alias{cf_account}
\title{Retrieve account information}
\usage{
cf_account(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A list containing account details for the authenticated user.
}
\description{
Retrieves CrowdFlower account information for the registered user.
related to the API key for the authenticated user
}
\examples{
\dontrun{
Sys.setenv("CROWDFLOWER_KEY" = "examplekey")
cf_account()
}
}
\references{
\href{https://success.crowdflower.com/hc/en-us/articles/202703425-CrowdFlower-API-Requests-Guide#account_information}{Crowdflower API documentation}
}
\seealso{
\code{\link{cf_jobs_list}}
}
\keyword{account}

