% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_channels.R
\name{job_channel_add}
\alias{job_channel_add}
\alias{job_channel_remove}
\title{Get and set job channels}
\usage{
job_channel_add(id, channel, verbose = TRUE, ...)

job_channel_remove(id, channel, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{channel}{A character string containing the name of a channel, as returned by \code{\link{job_channel_list}}.}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A logical \code{TRUE}, or an error.
}
\description{
Get, set, and replace the \dQuote{tags} for a job
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- job_create(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# list available channels
ch <- job_channel_list(j)
ch

# add new channel
job_channel_add(j, ch[1])

# remove a channel
job_channel_remove(j, ch[1])
}
}
\seealso{
\code{\link{job_create}}
}
\keyword{channels}

