% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/copy_job.R
\name{copyJob}
\alias{copyJob}
\title{Copy a job and, optionally, its data to a new job ID.}
\usage{
copyJob(id, rows = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{rows}{A logical indicating whether to copy rows from the original job.}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{APIcall}}.}
}
\value{
A character string contianing the job ID, invisibly.
}
\description{
\code{copyJob} copies an existing job to a new job ID.
}
\examples{
\dontrun{
copyJob(id = 'jobid')
}
}
\seealso{
\code{\link{createJob}}, \code{\link{updateJob}}
}

