% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_job.R
\name{createJob}
\alias{createJob}
\title{Creates a new job}
\usage{
createJob(title = NULL, instructions = NULL, cml = NULL)
}
\arguments{
\item{title}{Title for new job}

\item{instructions}{Instructions for new job}

\item{cml}{Layout for new job.}
}
\description{
\code{createJob} creates a new CrowdFlower Job.
}

