% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-account.R
\name{getAccount}
\alias{getAccount}
\title{Retrieve account information}
\usage{
getAccount(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link{APIcall}}.}
}
\value{
Account details.
}
\description{
\code{getAccount} retrieves the CrowdFlower account information
related to the API key for the authenticated user
}
\examples{
\dontrun{
getAccount()
}
}
\references{
\href{https://success.crowdflower.com/hc/en-us/articles/202703425-CrowdFlower-API-Requests-Guide#account_information}{Crowdflower API documentation}
}

