% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_jobs.R
\name{getJobs}
\alias{getJobs}
\title{Retrieve list of all jobs.}
\usage{
getJobs(n = 10, ...)
}
\arguments{
\item{n}{Number of recent jobs to return.}

\item{...}{Additional arguments passed to \code{\link{APIcall}}.}
}
\description{
\code{getJobs} queries for all jobs under the account related
to the API key for the authenticated user.
}

