% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jobs.R
\name{getJobs}
\alias{getJobs}
\title{Retrieve list of all jobs.}
\usage{
getJobs(page = NULL, ...)
}
\arguments{
\item{page}{A vector of integers specifying which page(s) of results 
to return. A page contains up to 10 jobs. The default (\code{NULL}) 
is to return all jobs.}

\item{...}{Additional arguments passed to \code{\link{APIcall}}.}
}
\value{
A data.frame containing details of all jobs. The \code{id} 
column provides the Crowdflower Job ID for each job.
}
\description{
\code{getJobs} queries for all jobs under the account related 
to the API key for the authenticated user.
}
\examples{
\dontrun{
# return first page of jobs
getJobs(page = 1)

# return all jobs
getJobs()
}

}
\references{
\href{https://success.crowdflower.com/hc/en-us/articles/202703425-CrowdFlower-API-Requests-Guide#account_information}{Crowdflower API documentation}
}
\seealso{
\code{\link{getAccount}}
}

