% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_status.R
\name{getStatus}
\alias{getStatus}
\title{Retrieve the status of a job}
\usage{
getStatus(id)
}
\arguments{
\item{id}{ID for job.}
}
\description{
\code{getStatus} gets the status of the job identified by the
id parameter of the request.
}

