% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_convert_gold.R
\name{job_convert_gold}
\alias{job_convert_gold}
\title{Convert uploaded gold for a job}
\usage{
job_convert_gold(id, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string specifying the ID for job whose gold questions will be converted.}

\item{verbose}{A logical indicating whether to print status information regarding the request.}

\item{\dots}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A logical.
}
\description{
Convert to test questions all rows where column \dQuote{_golden} is set to \dQuote{true}. Corresponds to the \dQuote{Convert Uploaded Test Questions} button on the Data page of the web interface.
}
\keyword{jobs}

