% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_convert_gold.R
\name{job_convert_gold}
\alias{job_convert_gold}
\title{Convert uploaded gold for a job}
\usage{
job_convert_gold(id, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string specifying the ID for job whose gold questions will be converted.}

\item{verbose}{A logical indicating whether to print status information regarding the request.}

\item{\dots}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A logical.
}
\description{
Convert to test questions all rows where column \dQuote{_golden} is set to \dQuote{true}. Corresponds to the \dQuote{Convert Uploaded Test Questions} button on the Data page of the web interface.
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- job_create(title = "Job Title", 
                instructions = readChar(f1, nchars = 1e8L),
                cml = readChar(f2, nchars = 1e8L))

# add data
d <- data.frame(content = c("hello", "goodbye", "world")
d[, "_golden"] <- c("true", "false", "false")
job_add_data(id = j, data = d)

# convert gold questions
job_convert_gold(id = j)

# launch job
job_launch(id = j)

# get results for job
report_regenerate(id = j, report_type = "full")
report_get(id = j, report_type = "full")

# delete job
job_delete(id = j)
}
}
\keyword{jobs}

