% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_delete.R
\name{job_delete}
\alias{job_delete}
\title{Delete a single job}
\usage{
job_delete(id, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A logical.
}
\description{
Delete a single Crowdflower job
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- job_create(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# delete job
job_delete(j)
}
}
\seealso{
\code{\link{job_create}}, \code{\link{job_update}}
}
\keyword{jobs}
