% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_update.R
\name{job_update}
\alias{job_update}
\title{Updates job settings.}
\usage{
job_update(id, title = NULL, instructions = NULL, cml = NULL,
  payment_cents = NULL, units_per_assignment = NULL, auto_launch = NULL,
  time_per_assignment = NULL, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job to be updated.}

\item{title}{New title.}

\item{instructions}{New set of instructions.}

\item{cml}{New layout for job.}

\item{payment_cents}{Amount, in U.S. cents, that contributors will be paid per task.}

\item{units_per_assignment}{Number of rows that will comprise one page within a job.}

\item{auto_launch}{A logical indicating whether to automatically launch rows as they are added.}

\item{time_per_assignment}{An integer specifying the length of time available per assignment, in seconds. Default is 1800 (i.e., 30 minutes).}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A character string containing the job ID, invisibly.
}
\description{
Updates settings for a given job.
}
\examples{
\dontrun{
# create a job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- job_create(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# update job's title
job_update(id, title = "New Title")

# update job's payment
job_update(id, payment_cents = "0.1")

# update job's instructions and units
job_update(id, instructions = "blank", units_per_assignment = 3)

# delete job
job_delete(j)
}
}
\seealso{
\code{\link{job_create}}
}
\keyword{jobs}
