% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{launch_job}
\alias{cancel_job}
\alias{launch_job}
\alias{pause_job}
\alias{resume_job}
\title{Control a job}
\usage{
launch_job(id, channel = "cf_internal", units = 100, verbose = FALSE, ...)

pause_job(id, verbose = FALSE, ...)

cancel_job(id, verbose = FALSE, ...)

resume_job(id, verbose = FALSE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{channel}{Either "on_demand" (for On Demand worforce), \dQuote{cf_internal} (for the Internal Channel) or "both"}

\item{units}{Units to be launched.}

\item{verbose}{A logical indicating whether to print details of the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
For \code{launch_job}, a list containing details of the job. Otherwise, a logical indicating the operation was successful, otherwise an error.
}
\description{
Launch, pause, resume, or cancel a job
}
\details{
\code{launch_job} launches a new job. \code{pause_job} pauses a currently running job. \code{resume_job} resumes a job. \code{cancel_job} cancels a job.
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- create_job(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# launch the job
launch_job(j)

# pause the job
pause_job(j)

# resume the job
resume_job(j)

# cancel the job
cancel_job(j)
}
}
\seealso{
\code{\link{pause_job}}, \code{\link{cancel_job}}, \code{\link{resume_job}}
}
\keyword{jobs}

