% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rows.R
\name{results_get}
\alias{results_get}
\title{Retrieve rows with judgment information for a job.}
\usage{
results_get(id, n = Inf, unit = NULL, type = c("aggregated", "full"),
  verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{n}{Number of rows to return. Specify only \code{n} or \code{unit}, but not both.}

\item{unit}{A \dQuote{unit id} for a specific row. Specify only \code{n} or \code{unit}, but not both.}

\item{type}{How responses should be aggregated ("aggregate", to take aggregated responses; or "full" for all responses)}

\item{verbose}{A logical indicating whether to print additional information about the request.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}
}
\value{
A data.frame containing judgment information
}
\description{
\code{results_get} retrieves rows with judgment information for the job ID indicated in the request. Specifying \code{unit} retrieves only judgments for a specific row (unit).
}
\examples{
\dontrun{
results_get('jobid')
}
}
\seealso{
\code{\link{job_status}}, \code{\link{report_get}}
}
\keyword{data}
\keyword{jobs}

