% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{resumeJob}
\alias{resumeJob}
\title{Pause a job}
\usage{
resumeJob(id, verbose = FALSE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{verbose}{A logical indicating whether to print details of the request.}

\item{...}{Additional arguments passed to \code{\link{APIcall}}.}
}
\value{
A logical indicating the operation was successful, otherwise an error.
}
\description{
\code{resumeJob} resumes a currently paused job.
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- createJob(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# launch the job
launchJob(j)

# pause the job
pauseJob(j)

# resume the job
resumeJob(j)

# cancel the job
cancelJob(j)
}

}
\seealso{
\code{\link{launchJob}}, \code{\link{cancelJob}}, \code{\link{pauseJob}}
}

