% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{getJobTags}
\alias{addJobTags}
\alias{getJobTags}
\alias{replaceJobTags}
\title{Get and set job tags}
\usage{
getJobTags(id, ...)

addJobTags(id, tags, verbose = TRUE, ...)

replaceJobTags(id, tags, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{...}{Additional arguments passed to \code{\link{APIcall}}.}

\item{tags}{For \code{addJobTags}, a character vector specifying one or more tags to add. For \code{replaceJobTags}, the same but to \emph{replace} rather than \emph{add} the tags.}

\item{verbose}{A logical indicating whether to print additional information about the request.}
}
\value{
For \code{getJobTags}, a character vector of tags. Otherwise, a logical \code{TRUE}, or an error.
}
\description{
Get, set, and replace the \dQuote{tags} for a job
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- createJob(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# get tags
getJobTags(j)

# add new tag
addJobTags(j, "textanalysis")

# replace tags
addJobTags(j, c("foo", "bar"))
getJobTags(j)
}

}
\seealso{
\code{\link{createJob}}
}

