% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_tags.R
\name{job_tags_get}
\alias{job_tags_add}
\alias{job_tags_get}
\alias{job_tags_replace}
\title{Get and set job tags}
\usage{
job_tags_get(id, ...)

job_tags_add(id, tags, verbose = TRUE, ...)

job_tags_replace(id, tags, verbose = TRUE, ...)
}
\arguments{
\item{id}{A character string containing an ID for job.}

\item{...}{Additional arguments passed to \code{\link{cf_query}}.}

\item{tags}{For \code{job_tags_add}, a character vector specifying one or more tags to add. For \code{job_tags_replace}, the same but to \emph{replace} rather than \emph{add} the tags.}

\item{verbose}{A logical indicating whether to print additional information about the request.}
}
\value{
For \code{job_tags_get}, a character vector of tags. Otherwise, a logical \code{TRUE}, or an error.
}
\description{
Get, set, and replace the \dQuote{tags} for a job
}
\examples{
\dontrun{
# create new job
f1 <- system.file("templates/instructions1.html", package = "crowdflower")
f2 <- system.file("templates/cml1.xml", package = "crowdflower")
j <- job_create(title = "Job Title", 
               instructions = readChar(f1, nchars = 1e8L),
               cml = readChar(f2, nchars = 1e8L))

# get tags
job_tags_get(j)

# add new tag
job_tags_add(j, "textanalysis")

# replace tags
job_tags_replace(j, c("foo", "bar"))
job_tags_get(j)

# delete job
job_delete(j)
}
}
\seealso{
\code{\link{job_create}}
}
\keyword{tags}

