% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update_job.R
\name{updateJob}
\alias{updateJob}
\title{Updates settings for a given job.}
\usage{
updateJob(id, title = NULL, instructions = NULL, cml = NULL, ...)
}
\arguments{
\item{id}{ID for job to be updated.}

\item{title}{New title.}

\item{instructions}{New set of instructions.}

\item{cml}{New layout for job.}

\item{...}{Additional arguments passed to \code{\link{APIcall}}.}
}
\description{
Updates settings for a given job.
}

